Extract `StorageProofChecker` utility from `bridges` to `substrate/frame`
WIP: squash

WIP: squash to primitives

WIP: squash to voting
