+++ bridges/primitives/runtime/Cargo.toml
@@ -20,12 +20,12 @@ scale-info = { features = ["derive"], workspace = true }
 serde = { features = ["alloc", "derive"], workspace = true }
 
 # Substrate Dependencies
+frame-proofs-primitives = { workspace = true }
 frame-support = { workspace = true }
 frame-system = { workspace = true }
 sp-core = { workspace = true }
 sp-io = { workspace = true }
 sp-runtime = { features = ["serde"], workspace = true }
-sp-state-machine = { workspace = true }
 sp-std = { workspace = true }
 sp-trie = { workspace = true }
 trie-db = { workspace = true }
@@ -37,6 +37,7 @@ hex-literal = { workspace = true, default-features = true }
 default = ["std"]
 std = [
 	"codec/std",
+	"frame-proofs-primitives/std",
 	"frame-support/std",
 	"frame-system/std",
 	"hash-db/std",
@@ -47,7 +48,6 @@ std = [
 	"sp-core/std",
 	"sp-io/std",
 	"sp-runtime/std",
-	"sp-state-machine/std",
 	"sp-std/std",
 	"sp-trie/std",
 	"trie-db/std",
