+++ substrate/frame/proofs/primitives/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "frame-proofs-primitives"
+version = "1.0.0"
+authors.workspace = true
+edition.workspace = true
+license = "Apache-2.0"
+homepage.workspace = true
+repository.workspace = true
+description = "Simple primitives for proofs and validation"
+
+[lints]
+workspace = true
+
+[dependencies]
+codec = { features = ["derive"], workspace = true }
+frame-support = { workspace = true }
+hash-db = { workspace = true }
+log = { workspace = true }
+scale-info = { features = ["derive"], workspace = true }
+sp-core = { workspace = true }
+sp-state-machine = { workspace = true }
+sp-runtime = { workspace = true }
+sp-trie = { workspace = true }
+
+[features]
+default = ["std"]
+std = [
+	"codec/std",
+	"frame-support/std",
+	"hash-db/std",
+	"log/std",
+	"scale-info/std",
+	"sp-core/std",
+	"sp-runtime/std",
+	"sp-state-machine/std",
+	"sp-trie/std",
+]
